% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_model-building.R
\name{add_covariates}
\alias{add_covariates}
\title{Add fixed effects of one or several covariates to some parameters.}
\usage{
add_covariates(nm, ..., use_regexpr = TRUE)
}
\arguments{
\item{nm}{A \code{networkModel} object.}

\item{...}{One or several formulas defining the covariates.}

\item{use_regexpr}{Boolean, use regular expression to match the parameters
affected by the formulas?}
}
\value{
A \code{networkModel} object.
}
\description{
New global parameters are given a default prior of hcauchy(scale = 1)
}
