% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_predictions.R
\name{as.mcmc.list.tidy_flows}
\alias{as.mcmc.list.tidy_flows}
\title{Convert a \code{tidy_flows} object to an \code{mcmc.list}}
\usage{
\method{as.mcmc.list}{tidy_flows}(x, ...)
}
\arguments{
\item{x}{A tidy flow object, as returned by \code{\link{tidy_flows}}. Note
that all chains must have the same iterations extracted (i.e. you must
use \code{n_per_chain} when calling \code{\link{tidy_flows}}).}

\item{...}{Not used for now.}
}
\value{
A \code{mcmc.list} object, with ordered iterations.
}
\description{
Convert a \code{tidy_flows} object to an \code{mcmc.list}
}
