% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_predictions.R
\name{as.mcmc.list.tidy_steady_states}
\alias{as.mcmc.list.tidy_steady_states}
\title{Convert a \code{tidy_steady_states} object to an \code{mcmc.list}}
\usage{
\method{as.mcmc.list}{tidy_steady_states}(x, ...)
}
\arguments{
\item{x}{A tidy steady states object, as returned by
\code{\link{tidy_steady_states}}. Note that all chains must have the
same iterations extracted (i.e. you must use \code{n_per_chain} when
calling \code{\link{tidy_flows}}).}

\item{...}{Not used for now.}
}
\value{
A \code{mcmc.list} object, with ordered iterations.
}
\description{
Convert a \code{tidy_steady_states} object to an \code{mcmc.list}
}
