% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_topology.R
\name{as_tbl_graph}
\alias{as_tbl_graph}
\title{Generic for as_tbl_graph()}
\usage{
as_tbl_graph(x, ...)
}
\arguments{
\item{x}{Object to convert to a tbl_graph.}

\item{...}{Passed to the appropriate method.}
}
\value{
A tbl_graph object.
}
\description{
Convert a compatible object to a tbl_graph object (from the tidygraph package)
}
