% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_topology.R
\name{as_tbl_graph.topology}
\alias{as_tbl_graph.topology}
\title{Convert a network topology to a tbl_graph}
\usage{
\method{as_tbl_graph}{topology}(x, ...)
}
\arguments{
\item{x}{A network topology.}

\item{...}{Not used.}
}
\value{
A tbl_graph object.
}
\description{
Convert a network topology to a tbl_graph
}
