% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{available_priors}
\alias{available_priors}
\title{List the available priors for model parameters}
\usage{
available_priors()
}
\value{
A tibble containing information about the available priors.
}
\description{
List the available priors for model parameters
}
\examples{
available_priors()

}
