% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_methods-mcmc.list.R
\name{c.mcmc.list}
\alias{c.mcmc.list}
\title{Combine mcmc.list objects}
\usage{
\method{c}{mcmc.list}(...)
}
\arguments{
\item{...}{\code{mcmc.list} objects.}
}
\value{
A \code{mcmc.list} object.
}
\description{
Combine mcmc.list objects
}
