% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_projection-simulation.R
\name{calculate_steady_state}
\alias{calculate_steady_state}
\title{Calculate steady-state compartment sizes for a network}
\usage{
calculate_steady_state(nm)
}
\arguments{
\item{nm}{A network model, with set parameter values.}
}
\value{
A tibble containing steady-state compartment sizes.
}
\description{
This is an experimental function. It attempts to calculate steady-state
compartment sizes using the set parameter values and the initial compartment
sizes. Use it with caution!
}
\details{
Note about how steady state sizes for split compartments are calculated: the
steady size of the active portion is calculated divide it is divided by the
active fraction (portion.act parameter) to get the total size including the
refractory portion. In this case we get a "steady-state" refractory portion,
consistent with steady state size of active fraction and with portion.act
parameter.
}
\examples{
m <- aquarium_mod
m <- set_prior(m, constant_p(0), "lambda")
m <- set_params(m, sample_params(m))
proj <- project(m, end = 40)
plot(proj)

z <- calculate_steady_state(m)
z
z$stable_sizes

}
