% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{constant}
\alias{constant}
\title{Define a fixed-value prior}
\usage{
constant(value)
}
\arguments{
\item{value}{The constant value of the parameter.}
}
\value{
A list defining the prior.
}
\description{
This is equivalent to having a fixed parameter.
}
\examples{
constant(2)

}
