% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{constant_p}
\alias{constant_p}
\title{Define a fixed-value prior}
\usage{
constant_p(value)
}
\arguments{
\item{value}{The constant value of the parameter.}
}
\value{
A list defining the prior.
}
\description{
This is equivalent to having a fixed parameter.
}
\examples{
constant_p(2)

}
