% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{exponential_p}
\alias{exponential_p}
\title{Define an exponential prior}
\usage{
exponential_p(lambda)
}
\arguments{
\item{lambda}{Lambda parameter (rate) of the exponential distribution. The
mean of the exponential distribution is 1/lambda.}
}
\value{
A list defining the prior.
}
\description{
Define an exponential prior
}
\examples{
exponential_p(0.5)

}
