% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_predictions.R
\name{filter.ppcNetworkModel}
\alias{filter.ppcNetworkModel}
\title{Filter method for output of tidy_data_and_posterior_predict()}
\usage{
\method{filter}{ppcNetworkModel}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A ppcNetworkModel object.}

\item{...}{Passed to dplyr::filter.}

\item{.preserve}{Ignored.}
}
\value{
A pccNetworkModel object filtered appropriately based on the
    [["vars"]] tibble.
}
\description{
Filter method for output of tidy_data_and_posterior_predict()
}
