% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_predictions.R
\name{filter}
\alias{filter}
\title{Filter (alias for filter function from dplyr)}
\usage{
filter(.data, ..., .by = NULL, .preserve = FALSE)
}
\arguments{
\item{.data}{Data to filter.}

\item{..., .by, .preserve}{Passed to dplyr::filter.}
}
\value{
See the returned value for dplyr::filter.
}
\description{
Filter (alias for filter function from dplyr)
}
