% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{format.prior}
\alias{format.prior}
\title{Pretty formatting of a \code{prior} object}
\usage{
\method{format}{prior}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prior}.}

\item{...}{Not used.}
}
\value{
A character string for pretty printing of a prior.
}
\description{
Pretty formatting of a \code{prior} object
}
