% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{format.prior_tibble}
\alias{format.prior_tibble}
\title{Pretty formatting of a \code{prior_tibble} object}
\usage{
\method{format}{prior_tibble}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prior_tibble}.}

\item{...}{Not used.}
}
\value{
A character string for pretty printing of a prior tibble.
}
\description{
Pretty formatting of a \code{prior_tibble} object
}
