% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_plotting-topology.R
\name{ggtopo.networkModel}
\alias{ggtopo.networkModel}
\title{Plot a network topology}
\usage{
\method{ggtopo}{networkModel}(x, layout = "auto", edge = "fan", ...)
}
\arguments{
\item{x}{A topology matrix.}

\item{layout}{Optional, layout to use (e.g. "sugiyama", "kk", "stress")}

\item{edge}{"curve" (the default) or "line".}

\item{...}{Not used for now.}
}
\value{
A ggplot2 plot.
}
\description{
A quick plot using ggraph
}
\examples{
if (requireNamespace("ggraph")) {
  ggtopo(aquarium_mod, edge = "line")
  ggtopo(trini_mod)
}

}
