% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_plotting-topology.R
\name{ggtopo}
\alias{ggtopo}
\title{Plot a topology}
\usage{
ggtopo(x, layout = "auto", edge = "fan", ...)
}
\arguments{
\item{x}{A network model or a topology matrix.}

\item{layout}{Optional, layout to use (e.g. "sugiyama", "kk", "stress")}

\item{edge}{"fan" (the default) or "line" or "curve".}

\item{...}{Passed to the methods.}
}
\value{
A ggplot2 plot.
}
\description{
A quick plot using ggraph
}
\examples{
if (requireNamespace("ggraph")) {
  ggtopo(aquarium_mod, edge = "line")
}

}
