% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_getters.R
\name{groups.networkModel}
\alias{groups.networkModel}
\title{Get the grouping for a \code{networkModel} object}
\usage{
\method{groups}{networkModel}(x)
}
\arguments{
\item{x}{A \code{networkModel} object.}
}
\value{
A tibble giving the grouping variable(s) for the input network
    model. This tibble is in the same order as the rows of the input network
    model. If the input network model did not have any grouping variable,
    returns \code{NULL}.
}
\description{
Get the grouping for a \code{networkModel} object
}
\examples{
groups(aquarium_mod)
groups(trini_mod)

}
