% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_methods-mcmc.list.R
\name{Math.mcmc.list}
\alias{Math.mcmc.list}
\title{Math generics for mcmc.list objects}
\usage{
\method{Math}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{\code{\link[coda]{mcmc.list}} object}

\item{...}{Other arguments passed to corresponding methods}
}
\value{
A \code{mcmc.list} object (with the added class
    \code{derived.mcmc.list}).
}
\description{
Math generics for mcmc.list objects
}
