% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_plotting.R
\name{mcmc_heatmap}
\alias{mcmc_heatmap}
\title{Draw a heatmap based on the correlations between parameters}
\usage{
mcmc_heatmap(x, col = NULL, ...)
}
\arguments{
\item{x}{A \code{coda::mcmc.list} object.}

\item{col}{Optional, vectors of colors defining the color ramp. Default uses
the divergent palette "Blue-Red 2" from the colorspace package.}

\item{...}{Passed to \code{\link{heatmap}}.}
}
\value{
Called for side effect (plotting).
}
\description{
Note that the colors represent the strength of the correlations (from 0 to
1), but do not inform about their sign. The method used to calculate
correlation coefficients is Spearman's rho.
}
