% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_model-building.R
\name{new_networkModel}
\alias{new_networkModel}
\title{Create an empty network model}
\usage{
new_networkModel(quiet = FALSE)
}
\arguments{
\item{quiet}{Boolean, if \code{FALSE} print a message indicating which
distribution family is used for proportions.}
}
\value{
An empty \code{networkModel} object. It is basically a zero-row
    tibble with the appropriate columns.
}
\description{
The first step in building a network model is to create a new, empty
\code{networkModel} object. This model can then be completed using functions
such as \code{set_topo()}, \code{set_init()}, etc...
}
\examples{
m <- new_networkModel()
m
class(m)

}
