% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{normal_p}
\alias{normal_p}
\title{Define a truncated normal prior (on [0;+Inf])}
\usage{
normal_p(mean, sd)
}
\arguments{
\item{mean}{Mean of the untruncated normal.}

\item{sd}{Standard deviation of the untruncated normal.}
}
\value{
A list defining the prior.
}
\description{
Define a truncated normal prior (on [0;+Inf])
}
\examples{
normal_p(mean = 0, sd = 4)

}
