% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_topology.R
\name{Ops.topology}
\alias{Ops.topology}
\title{Ops generics for \code{topology} objects}
\usage{
\method{Ops}{topology}(e1, e2)
}
\arguments{
\item{e1}{First operand}

\item{e2}{Second operand}
}
\value{
Boolean (or throws an error for unsupported operators).
}
\description{
Ops generics for \code{topology} objects
}
\examples{
topo(aquarium_mod) == topo(trini_mod)
topo(aquarium_mod) == topo(aquarium_mod)

}
