% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_getters.R
\name{params}
\alias{params}
\title{Return the parameters of a network model}
\usage{
params(nm, tibble = FALSE, expand = FALSE)
}
\arguments{
\item{nm}{A \code{networkModel} object.}

\item{tibble}{If TRUE, return a tidy tibble with the parameter mapping.}

\item{expand}{If TRUE, return a detailed tibble. The "tibble" argument is
disregarded in this case.}
}
\value{
A sorted character vector containing the parameters names.
}
\description{
Return the parameters of a network model
}
\examples{
params(aquarium_mod)
params(trini_mod)

}
