% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{pillar_shaft.prior}
\alias{pillar_shaft.prior}
\title{Function used for displaying \code{prior} object in tibbles}
\usage{
\method{pillar_shaft}{prior}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prior}.}

\item{...}{Not used.}
}
\value{
An object prepared with pillar::new_pillar_shaft_simple.
}
\description{
Function used for displaying \code{prior} object in tibbles
}
