% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_plotting.R
\name{plot.networkModel}
\alias{plot.networkModel}
\title{Plot observations/trajectories/predictions from a network model}
\usage{
\method{plot}{networkModel}(x, ...)
}
\arguments{
\item{x}{A \code{networkModel} object.}

\item{...}{Passed to \code{plot_nm}.}
}
\value{
Called for side effect (plotting).
}
\description{
Plot observations/trajectories/predictions from a network model
}
