% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_plotting.R
\name{plot.ready_for_unit_plot}
\alias{plot.ready_for_unit_plot}
\title{Plot output from \code{split_to_unit_plot}}
\usage{
\method{plot}{ready_for_unit_plot}(x, ...)
}
\arguments{
\item{x}{A \code{ready_for_unit_plot} object.}

\item{...}{Passed to \code{plot_nm}.}
}
\value{
Called for side effect (plotting).
}
\description{
Plot output from \code{split_to_unit_plot}
}
