% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_predictions.R
\name{posterior_predict.networkModelStanfit}
\alias{posterior_predict.networkModelStanfit}
\title{Draw from the posterior predictive distribution of the model outcome}
\usage{
\method{posterior_predict}{networkModelStanfit}(object, newdata, draw = NULL, cores = NULL, ...)
}
\arguments{
\item{object}{A networkModelStanfit object.}

\item{newdata}{Should be the model used to fit the networkStanfit object.}

\item{draw}{Integer, number of draws to perform from the posterior. Default
is 100.}

\item{cores}{Number of cores to use for parallel calculations. Default is
\code{NULL}, which means to use the value stored in
\code{options()[["mc.cores"]]} (or 1 if this value is not set).}

\item{...}{Not used for now.}
}
\value{
A "draw" by N matrix where N is the number of data points predicted
    per draw.
}
\description{
Draw from the posterior predictive distribution of the model outcome
}
