% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_predictions.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Draw from the posterior predictive distribution of the model outcome}
\usage{
posterior_predict(object, ...)
}
\arguments{
\item{object}{Model from which posterior predictions can be made.}

\item{...}{Passed to the appropriate method.}
}
\value{
Usually methods will implement a \code{draw} parameter, and the
    returned object is a "draw" by N matrix where N is the number of data
    points predicted per draw.
}
\description{
Draw from the posterior predictive distribution of the model outcome
}
