% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_model-building.R
\name{print.networkModel}
\alias{print.networkModel}
\title{Print method for \code{networkModel} objects}
\usage{
\method{print}{networkModel}(x, ...)
}
\arguments{
\item{x}{A \code{networkModel} object.}

\item{...}{Passsed to the next method.}
}
\value{
Called for the side effect of printing a network model object.
}
\description{
Print method for \code{networkModel} objects
}
