% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{print.prior}
\alias{print.prior}
\title{Pretty printing of a \code{prior} object}
\usage{
\method{print}{prior}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prior}.}

\item{...}{Not used.}
}
\value{
Mostly called for its side effect of printing, but also returns its
    input invisibly.
}
\description{
Pretty printing of a \code{prior} object
}
