% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{print.prior_tibble}
\alias{print.prior_tibble}
\title{Pretty printing of a \code{prior_tibble} object}
\usage{
\method{print}{prior_tibble}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prior_tibble}.}

\item{...}{Not used.}
}
\value{
Mostly called for its side effect of printing, but also returns its
    input invisibly.
}
\description{
Pretty printing of a \code{prior_tibble} object
}
