% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_topology.R
\name{print.topology}
\alias{print.topology}
\title{Pretty printing of a \code{topology} object}
\usage{
\method{print}{topology}(x, help = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{topology}.}

\item{help}{If TRUE, display a short help after the topology object
explaining e.g. the steady state or the split compartment symbols.}

\item{...}{Not used.}
}
\value{
Mostly called for its side effect (printing).
}
\description{
Pretty printing of a \code{topology} object
}
