% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_getters.R
\name{priors}
\alias{priors}
\title{Return the tibble containing the priors of a networkModel}
\usage{
priors(nm, fix_set_params = FALSE, quiet = FALSE)
}
\arguments{
\item{nm}{A \code{networkModel} object.}

\item{fix_set_params}{If TRUE, parameters for which a value is set are given a
fixed value (i.e. their prior is equivalent to a point value).}

\item{quiet}{Boolean to control verbosity.}
}
\value{
A tibble giving the current priors defined for the input network
    model.
}
\description{
Return the tibble containing the priors of a networkModel
}
\examples{
priors(aquarium_mod)
priors(trini_mod)

}
