% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_getters.R
\name{prop_family}
\alias{prop_family}
\title{Return the distribution family for observed proportions}
\usage{
prop_family(nm, quiet = FALSE)
}
\arguments{
\item{nm}{A \code{networkModel} object.}

\item{quiet}{Boolean for being quiet about explaining the role of eta
(default is \code{FALSE}).}
}
\value{
A character string describing the distribution family used to model
    observed proportions.
}
\description{
Return the distribution family for observed proportions
}
\examples{
prop_family(aquarium_mod)
prop_family(trini_mod)

}
