% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_plotting.R
\name{quick_sankey}
\alias{quick_sankey}
\title{Draw a Sankey plot with basic defaults}
\usage{
quick_sankey(flows, ...)
}
\arguments{
\item{flows}{A tibble containing flows (output from
\code{\link{tidy_flows}}). For now it should have an "average_flow"
column in the tibbles of the "flows" list column.}

\item{...}{Passed to \code{\link{sankey}}.}
}
\value{
Mostly called for its side effect (plotting), but also returns
    invisible the scene object describing the Sankey plot. Note that the
    structure of this object is experimental and might change in the future!
}
\description{
Draw a Sankey plot with basic defaults
}
