% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{groups}
\alias{select}
\alias{as.mcmc.list}
\alias{varnames}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coda}{\code{\link[coda:as.mcmc.list]{as.mcmc.list()}}, \code{\link[coda:varnames]{varnames()}}}

  \item{dplyr}{\code{\link[dplyr:groups]{groups()}}, \code{\link[dplyr:select]{select()}}}

  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}
}}

