% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_predictions.R
\name{sample_params}
\alias{sample_params}
\title{Sample parameter values from priors}
\usage{
sample_params(nm)
}
\arguments{
\item{nm}{A \code{networkModel} object.}
}
\value{
A named vector containing parameter values.
}
\description{
Sample parameter values from priors
}
\examples{
library(magrittr)

p <- sample_params(aquarium_mod)
p

proj <- aquarium_mod \%>\% set_params(p) \%>\% project(end = 10)
plot(proj)

}
