% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{scaled_beta_p}
\alias{scaled_beta_p}
\title{Define a beta prior (on [0;scale])}
\usage{
scaled_beta_p(alpha, beta, scale = 1)
}
\arguments{
\item{alpha}{Alpha parameter of the unscaled beta distribution.}

\item{beta}{Beta parameter of the unscaled beta distribution.}

\item{scale}{The upper boundary of the prior.}
}
\value{
A list defining the prior.
}
\description{
If a random variable X follows a scaled beta distribution with parameters
(alpha, beta, scale), then X/scale follows a beta distribution with
parameters (alpha, beta).
}
\examples{
scaled_beta_p(0.8, 20, scale = 10)

}
