% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_methods-mcmc.list.R
\name{select.mcmc.list}
\alias{select.mcmc.list}
\title{Select parameters based on their names}
\usage{
\method{select}{mcmc.list}(.data, ...)
}
\arguments{
\item{.data}{A \code{coda::mcmc.list} object.}

\item{...}{Strings used to select variables using pattern matching with
\code{grepl}.}
}
\value{
An \code{mcmc.list} object, with the same extra class(es) as
    \code{.data} (if any).
}
\description{
Select parameters based on their names
}
