% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_model-building.R
\name{set_init}
\alias{set_init}
\title{Set initial conditions in a network model}
\usage{
set_init(nm, data, comp, size, prop, group_by = NULL)
}
\arguments{
\item{nm}{A \code{networkModel} object (e.g. output from
\code{\link{new_networkModel}})}

\item{data}{A tibble containing the initial conditions}

\item{comp}{String, name of the \code{data} column with the compartment names}

\item{size}{String, name of the \code{data} column with the compartment sizes}

\item{prop}{String, name of the \code{data} column with the compartment
proportions of marked tracer}

\item{group_by}{Optional vector of string giving the names of the columns to
use for grouping the data into replicates}
}
\value{
A \code{networkModel} object.
}
\description{
Set initial conditions in a network model
}
