% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_model-building.R
\name{set_obs}
\alias{set_obs}
\title{Set observations in a network model}
\usage{
set_obs(nm, data, comp, size, prop, time, group_by)
}
\arguments{
\item{nm}{A \code{networkModel} object (e.g. output from
\code{\link{new_networkModel}})}

\item{data}{A tibble containing the observations. If NULL, remove
observations from the model.}

\item{comp}{String, name of the \code{data} column with the compartment
names}

\item{size}{String, name of the \code{data} column with the compartment
sizes}

\item{prop}{String, name of the \code{data} column with the compartment
proportions of heavy tracer}

\item{time}{String, name of the \code{data} column with the sampling times}

\item{group_by}{Optional vector of string giving the names of the columns to
use for grouping the data into replicates}
}
\value{
A \code{networkModel} object.
}
\description{
Set observations in a network model
}
