% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_model-building.R
\name{set_prior}
\alias{set_prior}
\title{Set prior(s) for a network model}
\usage{
set_prior(x, prior, param, use_regexp = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{A \code{networkModel} object.}

\item{prior}{A prior built with e.g. uniform() or hcauchy().}

\item{param}{String, target parameter or regexp to target several parameters}

\item{use_regexp}{Boolean, if \code{TRUE} (the default) then \code{param} is
used as a regular expression to match one or several parameter names.}

\item{quiet}{Boolean, if \code{FALSE} print a message indicating which
parameters had their prior modified.}
}
\value{
A \code{networkModel} object.
}
\description{
Set prior(s) for a network model
}
