% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_getters.R
\name{size_family}
\alias{size_family}
\title{Return the distribution family for observed sizes}
\usage{
size_family(nm, quiet = FALSE)
}
\arguments{
\item{nm}{A \code{networkModel} object.}

\item{quiet}{Boolean for being quiet about explaining the role of zeta
(default is \code{FALSE}).}
}
\value{
A character string describing the distribution family used to model
    observed sizes.
}
\description{
Return the distribution family for observed sizes
}
\examples{
size_family(aquarium_mod)
size_family(trini_mod)

}
