% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_run-mcmc.R
\name{stanfit_to_named_mcmclist}
\alias{stanfit_to_named_mcmclist}
\title{Convert a Stanfit object to a nicely named mcmc.list object}
\usage{
stanfit_to_named_mcmclist(stanfit)
}
\arguments{
\item{stanfit}{A stanfit object returned by \code{rstan::sampling}.}
}
\value{
An \code{mcmc.list} object. It also has the original stanfit object
  stored as an attribute \code{"stanfit"}.
}
\description{
When running \code{run_mcmc} with \code{stanfit = FALSE} (typically for
debugging purposes), the parameters in the returned \code{stanfit} object
are named using a base label and an indexing system. This function provides
a way to convert this \code{stanfit} object into a more conventional
\code{mcmc.list} object where parameters are named according to their role
in the original network model used when running \code{run_mcmc}.
}
