% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_methods-mcmc.list.R
\name{[.networkModelStanfit}
\alias{[.networkModelStanfit}
\title{Subset method for \code{networkModelStanfit} objects}
\usage{
\method{[}{networkModelStanfit}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{A \code{networkModelStanfit} object.}

\item{i}{A vector of iteration indices.}

\item{j}{A vector of parameter names or indices.}

\item{drop}{Boolean.}
}
\value{
A \code{networkModelStanfit} object.
}
\description{
Subset method for \code{networkModelStanfit} objects
}
