% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_plotting.R
\name{traceplot}
\alias{traceplot}
\title{Plot mcmc.list objects}
\usage{
traceplot(x, ...)
}
\arguments{
\item{x}{A \code{coda::mcmc.list} object.}

\item{...}{Passed to \code{plot_traces}.}
}
\value{
Called for side effect (plotting).
}
\description{
Plot mcmc.list objects
}
