% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{type_sum.prior}
\alias{type_sum.prior}
\title{Function used for displaying \code{prior} object in tibbles}
\usage{
\method{type_sum}{prior}(x)
}
\arguments{
\item{x}{An object of class \code{prior}.}
}
\value{
Input formatted with \code{format(x)}.
}
\description{
Function used for displaying \code{prior} object in tibbles
}
