% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{uniform_p}
\alias{uniform_p}
\title{Define a uniform prior}
\usage{
uniform_p(min, max)
}
\arguments{
\item{min, max}{Minimum and maximum boundaries for the uniform prior.}
}
\value{
A list defining the prior.
}
\description{
Define a uniform prior
}
\examples{
uniform_p(min = 0, max= 1)

}
