% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_predictions.R
\name{filter}
\alias{filter}
\title{Filter (alias for filter function from dplyr)}
\arguments{
\item{.data}{Data to filter.}

\item{...}{Passed to dplyr::filter.}

\item{preserve}{Ignored.}
}
\value{
See the returned value for dplyr::filter.
}
\description{
Filter (alias for filter function from dplyr)
}
