% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{isotracer-package}
\alias{isotracer-package}
\alias{isotracer}
\title{The 'isotracer' package}
\description{
The isotracer package allows modelling of fluxes across a network of
compartments. Parameters are estimated using a Bayesian MCMC approach.
}
\references{
López-Sepulcre, A., M. Bruneaux, S. M. Collins, R. El-Sabaawi,
A. S. Flecker, and S. A. Thomas. The American Naturalist
(2020). "A New Method to Reconstruct Quantitative Food Webs and
Nutrient Flows from Isotope Tracer Addition Experiments."
https://doi.org/10.1086/708546.

Stan Development Team (2018). RStan: the R interface to Stan. R package
version 2.18.2. https://mc-stan.org
}
