% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{isotracer-package}
\alias{isotracer-package}
\alias{isotracer}
\title{The 'isotracer' package}
\description{
The isotracer package allows modelling of fluxes across a network of
compartments. Parameters are estimated using a Bayesian MCMC approach.
}
\references{
López-Sepulcre, A., M. Bruneaux, S. M. Collins, R. El-Sabaawi,
A. S. Flecker, and S. A. Thomas. The American Naturalist
(2020). "A New Method to Reconstruct Quantitative Food Webs and
Nutrient Flows from Isotope Tracer Addition Experiments."
https://doi.org/10.1086/708546.

Stan Development Team (2018). RStan: the R interface to Stan. R package
version 2.18.2. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/matthieu-bruneaux/isotracer}
  \item Report bugs at \url{https://gitlab.com/matthieu-bruneaux/isotracer/-/issues}
}

}
\author{
\strong{Maintainer}: Matthieu Bruneaux \email{matthieu.bruneaux@gmail.com} (\href{https://orcid.org/0000-0001-6997-192X}{ORCID})

Authors:
\itemize{
  \item Andrés López-Sepulcre \email{lopezsepulcre@gmail.com} (\href{https://orcid.org/0000-0001-9708-0788}{ORCID})
}

}
