% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_priors.R
\name{Ops.prior}
\alias{Ops.prior}
\title{Implementation of the '==' operator for priors}
\usage{
\method{Ops}{prior}(e1, e2)
}
\arguments{
\item{e1, e2}{Objects of class "prior".}
}
\value{
Boolean (or throws an error for unsupported operators).
}
\description{
Implementation of the '==' operator for priors
}
\examples{
p <- constant_p(0)
q <- constant_p(4)
p == q

p <- hcauchy_p(2)
q <- hcauchy_p(2)
p == q

}
