% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_getters.R
\name{params}
\alias{params}
\title{Return the parameters of a network model}
\usage{
params(nm, simplify = FALSE)
}
\arguments{
\item{nm}{A \code{networkModel} object.}

\item{simplify}{If \code{TRUE}, return a vector containing the names of all
model parameters (default: \code{FALSE}).}
}
\value{
A tibble containing the parameter names and their current value (if
    set). If \code{simplify} is \code{TRUE}, only return a sorted character
    vector containing the parameters names.
}
\description{
Return the parameters of a network model
}
\examples{
params(aquarium_mod)
params(trini_mod)
params(trini_mod, simplify = TRUE)

}
